разбил компонент карточек на отдельные компоненты Card_list и Card_item. Данные по рендеру переместил в папку data
