+++ src/data/dataCards.js
@@ -1,0 +1,52 @@
+export const dataCards = [
+  {
+    title: 'Card 1',
+    description: 'This is the description for card 1.',
+    imageUrl: 'https://robohash.org/Johnathon.png?set=set1',
+  },
+  {
+    title: 'Card 2',
+    description: 'This is the description for card 2.',
+    imageUrl: 'https://robohash.org/Johnathon.png?set=set2',
+  },
+  {
+    title: 'Card 3',
+    description: 'This is the description for card 3.',
+    imageUrl: 'https://robohash.org/Johnathon.png?set=set3',
+  },
+  {
+    title: 'Card 4',
+    description: 'This is the description for card 4.',
+    imageUrl: 'https://robohash.org/Johnathon.png?set=set4',
+  },
+  {
+    title: 'Card 5',
+    description: 'This is the description for card 5.',
+    imageUrl: 'https://robohash.org/Johnathon.png?set=set5',
+  },
+  {
+    title: 'Card 6',
+    description: 'This is the description for card 6.',
+    imageUrl: 'https://robohash.org/Johnathon.png?set=set1',
+  },
+  {
+    title: 'Card 7',
+    description: 'This is the description for card 7.',
+    imageUrl: 'https://robohash.org/Johnathon.png?set=set2',
+  },
+  {
+    title: 'Card 8',
+    description: 'This is the description for card 8.',
+    imageUrl: 'https://robohash.org/Johnathon.png?set=set3',
+  },
+  {
+    title: 'Card 9',
+    description: 'This is the description for card 9.',
+    imageUrl: 'https://robohash.org/Johnathon.png?set=set4',
+  },
+  {
+    title: 'Card 10',
+    description: 'This is the description for card 10.',
+    imageUrl: 'https://robohash.org/Johnathon.png?set=set5',
+  },
+];
